Fighting Implemented For the Prototype
Units have a "MyEnemyTag" String. This must be what Tag their enemies have (Player or Enemy).
They have a damage and life value.
If one Object with the appropriate enemy Tag enters the "EnemyDetector" that objet is added to a List. Then they do damage to the first Object in the List.